requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
opencv-python-headless>=4.8.0
mlflow>=2.8.0
Pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.65.0
fastapi>=0.100.0
uvicorn>=0.23.0
#추가
python-multipart